Merge the Working and Technical Committees
This merges the Working and Technical Committees into a single body
with 9 members, retaining the name "Technical Committee".  The new
TC has the responsibilities of both previous groups, and splits
the difference in election timing with an election that falls in July.

Feedback is welcome on this governance change. The current Working
Committee will ultimately vote, with four members' +2s required
to commit the change.

Change-Id: I4f000e3eef45a46ad09cd30f12868b3825e729f2